Migrate IDOL frontend to next.js
