--- frontend/.env
@@ -1,1 +1,0 @@
-SKIP_PREFLIGHT_CHECK=true
