+++ frontend/package.json
@@ -13,25 +13,23 @@
     "@types/react": "^17.0.24",
     "@types/react-avatar-editor": "^10.3.6",
     "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
     "axios": "^0.21.4",
     "firebase": "^7.19.1",
+    "next": "^11.1.2",
     "prismjs": "^1.25.0",
     "react": "^17.0.2",
     "react-avatar-editor": "^11.1.0",
     "react-dom": "^17.0.2",
     "react-markdown": "^5.0.3",
-    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^1.2.1",
     "typescript": "^4.4.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "tsc": "tsc",
-    "eject": "react-scripts eject"
+    "start": "next",
+    "build": "next build && next export",
+    "tsc": "tsc"
   },
   "proxy": "http://localhost:9000",
   "browserslist": {
