+++ frontend/src/components/Common/SiteHeader.module.css
@@ -1,5 +1,3 @@
-@import '../../variables.css';
-
 .SiteHeader {
   background: whitesmoke;
   display: flex;
