+++ frontend/src/firebase.ts
@@ -24,7 +24,8 @@ const firebaseConfig = useProdDb
       measurementId: 'G-2QB5YJ3CHC'
     };
 
-firebase.initializeApp(firebaseConfig);
+if (!firebase.apps.length) firebase.initializeApp(firebaseConfig);
+
 export const auth = firebase.auth();
 export const firestore = firebase.firestore();
 export const provider = new firebase.auth.GoogleAuthProvider();
