+++ frontend/src/pages/admin/edit.tsx
@@ -1,0 +1,3 @@
+import AddUser from '../../components/Admin/AddUser';
+
+export default AddUser;
