+++ frontend/src/pages/admin/member-review.tsx
@@ -1,0 +1,3 @@
+import MemberReview from '../../components/Admin/MemberReview';
+
+export default MemberReview;
