+++ frontend/src/pages/admin/shoutouts.tsx
@@ -1,0 +1,3 @@
+import AdminShoutouts from '../../components/Admin/AdminShoutouts';
+
+export default AdminShoutouts;
