+++ frontend/src/pages/admin/signin-creator.tsx
@@ -1,0 +1,3 @@
+import SignInFormCreatorBase from '../../components/Admin/SignInFormCreator';
+
+export default SignInFormCreatorBase;
