+++ frontend/src/pages/admin/site-deployer.tsx
@@ -1,0 +1,3 @@
+import SiteDeployer from '../../components/Admin/SiteDeployer';
+
+export default SiteDeployer;
