+++ frontend/src/pages/admin/teams/edit.tsx
@@ -1,0 +1,3 @@
+import EditTeam from '../../../components/Admin/EditTeam';
+
+export default EditTeam;
