+++ frontend/src/pages/forms/index.tsx
@@ -1,0 +1,23 @@
+import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard';
+
+const navCardItems: readonly NavigationCardItem[] = [
+  { header: 'Sign-In Form', description: 'Sign in to an event!', link: '/forms/signin' },
+  {
+    header: 'Shoutouts',
+    description: 'Give someone a shoutout or view your past given and received shoutouts.',
+    link: '/forms/shoutouts'
+  },
+  {
+    header: 'Edit Profile',
+    description: "Edit your profile information on DTI's website.",
+    link: '/forms/profile'
+  },
+  {
+    header: 'Edit Profile Image',
+    description: 'Edit your profile image.',
+    link: '/forms/profileImage'
+  }
+];
+
+const FormsIndex = (): JSX.Element => <NavigationCard items={navCardItems} />;
+export default FormsIndex;
