+++ frontend/src/pages/forms/profile.tsx
@@ -1,0 +1,3 @@
+import Profile from '../../components/Forms/UserProfile/UserProfile';
+
+export default Profile;
