+++ frontend/src/pages/forms/profileImage.tsx
@@ -1,0 +1,3 @@
+import ProfileImage from '../../components/Forms/UserProfile/UserProfileImage';
+
+export default ProfileImage;
