+++ frontend/src/pages/forms/shoutouts.tsx
@@ -1,0 +1,3 @@
+import ShoutoutsPage from '../../components/Forms/ShoutoutsPage/ShoutoutsPage';
+
+export default ShoutoutsPage;
