+++ frontend/src/pages/forms/signin.tsx
@@ -1,0 +1,3 @@
+import SignInForm from '../../components/Forms/SignInForm/SignInForm';
+
+export default SignInForm;
