+++ frontend/src/pages/games/dti48.tsx
@@ -1,0 +1,3 @@
+import DTI48 from '../../components/Games/DTI48/DTI48';
+
+export default DTI48;
