+++ frontend/src/pages/games/index.tsx
@@ -1,0 +1,8 @@
+import NavigationCard, { NavigationCardItem } from '../../components/Common/NavigationCard';
+
+const navCardItems: readonly NavigationCardItem[] = [
+  { header: 'DTI48', description: 'Keep merging until you get Ashneel.', link: '/games/dti48' }
+];
+
+const GamesIndex = (): JSX.Element => <NavigationCard items={navCardItems} />;
+export default GamesIndex;
