+++ frontend/src/pages/index.tsx
@@ -1,0 +1,3 @@
+import Homepage from '../components/Homepage/Homepage.lazy';
+
+export default Homepage;
