--- frontend/src/react-app-env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="react-scripts" />
-/// <reference types="common-types" />
