+++ frontend/src/types.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="common-types" />
